+++ gui/theme_maker.go
@@ -375,7 +375,7 @@ func ThemeMaker(cfg ThemeCfg) Theme {
 		toastStyle: ToastStyle{
 			Shadow:       cfg.ShadowPopover,
 			maxVisible:   5,
-			Anchor:       toastBottomRight,
+			Anchor:       ToastBottomRight,
 			Width:        260,
 			margin:       16,
 			Spacing:      cfg.SpacingMedium,